DART build
quickbuild.sh
Only building the three programs filter, pmo, fill_inflation_restart
Removing Makefile at the end of quickbuild.sh to not have the makefile lying around

Removed the check for the case root input.nml for dart - not using this in the build, using preprocess_inpuit.nml from DART_interface
Note in buildnml why is input.nml getting copied to the caseroot? Is this just for the build or is this for running dart?

Builds and cleans successfuly ./case.build —build|clean esp
Clean is using the CIME tools Makefile:cleanesp section